Move all DAO classes as inner classes
