+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/db/NoteDao.java
@@ -3,8 +3,9 @@ package no.sikt.nva.nvi.common.db;
 import static no.sikt.nva.nvi.common.DatabaseConstants.DATA_FIELD;
 import static no.sikt.nva.nvi.common.DatabaseConstants.HASH_KEY;
 import static no.sikt.nva.nvi.common.DatabaseConstants.SORT_KEY;
+import java.time.Instant;
 import java.util.UUID;
-import no.sikt.nva.nvi.common.db.model.DbNote;
+import no.sikt.nva.nvi.common.db.model.Username;
 import nva.commons.core.JacocoGenerated;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbAttribute;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbImmutable;
@@ -86,6 +87,53 @@ public record NoteDao(UUID identifier,
 
         public NoteDao build() {
             return new NoteDao(builderIdentifier, builderNote);
+        }
+    }
+
+    @DynamoDbImmutable(builder = DbNote.Builder.class)
+    public record DbNote(UUID noteId,
+                         Username user,
+                         String text,
+                         Instant createdDate) {
+
+        public static Builder builder() {
+            return new Builder();
+        }
+
+        @JacocoGenerated //TODO use later :)
+        public static final class Builder {
+
+            private UUID builderNoteId;
+            private Username builderUser;
+            private String builderText;
+            private Instant builderCreatedDate;
+
+            public Builder() {
+            }
+
+            public Builder noteId(UUID noteId) {
+                this.builderNoteId = noteId;
+                return this;
+            }
+
+            public Builder user(Username user) {
+                this.builderUser = user;
+                return this;
+            }
+
+            public Builder text(String text) {
+                this.builderText = text;
+                return this;
+            }
+
+            public Builder createdDate(Instant createdDate) {
+                this.builderCreatedDate = createdDate;
+                return this;
+            }
+
+            public DbNote build() {
+                return new DbNote(builderNoteId, builderUser, builderText, builderCreatedDate);
+            }
         }
     }
 }
