+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/UpdateAssigneeRequest.java
@@ -1,6 +1,7 @@
 package no.sikt.nva.nvi.common.model;
 
-import no.sikt.nva.nvi.common.db.model.DbUsername;
+import no.sikt.nva.nvi.common.db.model.Username;
 
-public record UpdateAssigneeRequest(DbUsername username) implements UpdateApprovalRequest {
+public record UpdateAssigneeRequest(Username username) implements UpdateApprovalRequest {
+
 }
