+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/UpdateStatusRequest.java
@@ -1,6 +1,6 @@
 package no.sikt.nva.nvi.common.model;
 
-import no.sikt.nva.nvi.common.db.model.DbStatus;
+import no.sikt.nva.nvi.common.db.ApprovalStatusDao.DbStatus;
 
 public record UpdateStatusRequest(DbStatus approvalStatus,
                                   String username,
