+++ nvi-rest/src/main/java/no/sikt/nva/nvi/rest/create/CreateNoteHandler.java
@@ -7,8 +7,8 @@ import com.amazonaws.services.lambda.runtime.Context;
 import java.net.HttpURLConnection;
 import java.util.Objects;
 import java.util.UUID;
-import no.sikt.nva.nvi.common.db.model.DbNote;
-import no.sikt.nva.nvi.common.db.model.DbUsername;
+import no.sikt.nva.nvi.common.db.NoteDao.DbNote;
+import no.sikt.nva.nvi.common.db.model.Username;
 import no.sikt.nva.nvi.common.service.NviService;
 import no.sikt.nva.nvi.rest.model.CandidateResponse;
 import no.sikt.nva.nvi.rest.model.CandidateResponseMapper;
@@ -54,10 +54,10 @@ public class CreateNoteHandler extends ApiGatewayHandler<NviNoteRequest, Candida
         return HttpURLConnection.HTTP_OK;
     }
 
-    private static DbNote getNote(NviNoteRequest input, DbUsername username) {
+    private static DbNote getNote(NviNoteRequest input, Username username) {
         return DbNote.builder()
                    .text(input.text())
-                   .user(DbUsername.fromString(username.value()))
+                   .user(Username.fromString(username.value()))
                    .build();
     }
