+++ nvi-rest/src/main/java/no/sikt/nva/nvi/rest/create/CreateNviPeriodHandler.java
@@ -4,12 +4,12 @@ import static no.sikt.nva.nvi.utils.RequestUtil.getUsername;
 import static nva.commons.core.attempt.Try.attempt;
 import com.amazonaws.services.lambda.runtime.Context;
 import java.net.HttpURLConnection;
-import no.sikt.nva.nvi.utils.RequestUtil;
-import no.sikt.nva.nvi.rest.model.NviPeriodDto;
-import no.sikt.nva.nvi.common.db.model.DbNviPeriod;
-import no.sikt.nva.nvi.common.db.model.DbNviPeriod.Builder;
+import no.sikt.nva.nvi.common.db.NviPeriodDao.DbNviPeriod;
+import no.sikt.nva.nvi.common.db.NviPeriodDao.DbNviPeriod.Builder;
 import no.sikt.nva.nvi.common.service.NviService;
+import no.sikt.nva.nvi.rest.model.NviPeriodDto;
 import no.sikt.nva.nvi.utils.ExceptionMapper;
+import no.sikt.nva.nvi.utils.RequestUtil;
 import nva.commons.apigateway.AccessRight;
 import nva.commons.apigateway.ApiGatewayHandler;
 import nva.commons.apigateway.RequestInfo;
