+++ nvi-rest/src/main/java/no/sikt/nva/nvi/rest/fetch/FetchNviPeriodsHandler.java
@@ -4,7 +4,7 @@ import static nva.commons.core.attempt.Try.attempt;
 import com.amazonaws.services.lambda.runtime.Context;
 import java.net.HttpURLConnection;
 import java.util.List;
-import no.sikt.nva.nvi.common.db.model.DbNviPeriod;
+import no.sikt.nva.nvi.common.db.NviPeriodDao.DbNviPeriod;
 import no.sikt.nva.nvi.common.service.NviService;
 import no.sikt.nva.nvi.rest.model.NviPeriodDto;
 import no.sikt.nva.nvi.rest.model.NviPeriodsResponse;
@@ -37,12 +37,13 @@ public class FetchNviPeriodsHandler extends ApiGatewayHandler<Void, NviPeriodsRe
         RequestUtil.hasAccessRight(requestInfo, AccessRight.MANAGE_NVI_PERIODS);
 
         return attempt(nviService::getPeriods)
-                    .map(this::toNviPeriodsResponse)
-                    .orElseThrow();
+                   .map(this::toNviPeriodsResponse)
+                   .orElseThrow();
     }
 
-    private NviPeriodsResponse toNviPeriodsResponse(List<DbNviPeriod> dbNviPeriods) {
-        return new NviPeriodsResponse(toPeriodDtoList(dbNviPeriods));
+    @Override
+    protected Integer getSuccessStatusCode(Void input, NviPeriodsResponse output) {
+        return HttpURLConnection.HTTP_OK;
     }
 
     private static List<NviPeriodDto> toPeriodDtoList(List<DbNviPeriod> dbNviPeriods) {
@@ -51,9 +52,7 @@ public class FetchNviPeriodsHandler extends ApiGatewayHandler<Void, NviPeriodsRe
                    .toList();
     }
 
-    @Override
-    protected Integer getSuccessStatusCode(Void input, NviPeriodsResponse output) {
-        return HttpURLConnection.HTTP_OK;
+    private NviPeriodsResponse toNviPeriodsResponse(List<DbNviPeriod> dbNviPeriods) {
+        return new NviPeriodsResponse(toPeriodDtoList(dbNviPeriods));
     }
-
 }
