+++ nvi-rest/src/main/java/no/sikt/nva/nvi/rest/model/NviPeriodDto.java
@@ -2,12 +2,16 @@ package no.sikt.nva.nvi.rest.model;
 
 import static nva.commons.core.attempt.Try.attempt;
 import java.time.Instant;
-import no.sikt.nva.nvi.common.db.model.DbNviPeriod;
+import no.sikt.nva.nvi.common.db.NviPeriodDao.DbNviPeriod;
 
 public record NviPeriodDto(String publishingYear,
                            String reportingDate) {
 
     public static final String INVALID_REPORTING_DATE_MESSAGE = "Invalid reporting date";
+
+    public static NviPeriodDto fromNviPeriod(DbNviPeriod period) {
+        return new NviPeriodDto(period.publishingYear(), period.reportingDate().toString());
+    }
 
     public DbNviPeriod toNviPeriod() {
         return DbNviPeriod.builder()
@@ -20,9 +24,4 @@ public record NviPeriodDto(String publishingYear,
         return attempt(() -> Instant.parse(reportingDate))
                    .orElseThrow(failure -> new IllegalArgumentException(INVALID_REPORTING_DATE_MESSAGE));
     }
-
-    public static NviPeriodDto fromNviPeriod(DbNviPeriod period) {
-        return new NviPeriodDto(period.publishingYear(), period.reportingDate().toString());
-    }
-
 }
