+++ nvi-rest/src/main/java/no/sikt/nva/nvi/utils/RequestUtil.java
@@ -1,6 +1,6 @@
 package no.sikt.nva.nvi.utils;
 
-import no.sikt.nva.nvi.common.db.model.DbUsername;
+import no.sikt.nva.nvi.common.db.model.Username;
 import nva.commons.apigateway.AccessRight;
 import nva.commons.apigateway.RequestInfo;
 import nva.commons.apigateway.exceptions.UnauthorizedException;
@@ -12,8 +12,8 @@ public final class RequestUtil {
     private RequestUtil() {
     }
 
-    public static DbUsername getUsername(RequestInfo requestInfo) throws UnauthorizedException {
-        return DbUsername.fromString(requestInfo.getUserName());
+    public static Username getUsername(RequestInfo requestInfo) throws UnauthorizedException {
+        return Username.fromString(requestInfo.getUserName());
     }
 
     public static void hasAccessRight(RequestInfo requestInfo, AccessRight accessRight) throws UnauthorizedException {
